refactor: rename page component folders
